Update DbContextExtensions.cs
Made EntityEntry State to UnChanged. As soon as one of the property is modified, the state of the entity will also move to Modified. But only the properties we marked as modified will be updated when we save. I think this line was missing. Correct me if I am wrong!